Loop for test (fixing #9) - add loop for test f2 (20% of errors) - change show boolean

+++ src/coc_testinghall/testJSON.py
@@ -10,25 +10,19 @@ from Class.Satellite import Satellite
 from Class.Orbit import Orbit
 from src.coc_testinghall.TestingObject import Testing_system, Testing_Star, Testing_Planet
 import os
+import numpy as np
+
 os.chdir("../SSG")
 path = "/Users/floriandelrieu/OneDrive/Logiciels et Jeux/SpaceColony/src/coc_testinghall/"
 NonJSONWritableList = [Planet,Star,Satellite,Orbit,System,
                        Testing_system,Testing_Star,Testing_Planet]
 
-S = System()
-S.createPlanet()
-
-#S = Testing_system()
-#s1 = Testing_Star()
-#p1 = Testing_Planet()
-#s1.Orbit_list = [p1]
-#S.StarList = [s1]
-
-
 def SaveInJSON(Objet,json_name):
-    #dico = Planet.__dict__
-    #dico = f1(dico)
-    json_file = open(json_name+".json","w")
+    import os
+    if os.path.isfile(json_name):
+        json_file = open(json_name+".json","a")
+    else:
+        json_file = open(json_name+".json","w")
     json.dump(Objet, json_file)
 
 def f2(obj):
@@ -41,12 +35,12 @@ def f2(obj):
           - when obj=Planet() and el1="Planet"
             - why ?
     """
-
     if type(obj) in NonJSONWritableList:  # Verifie si :obj: est pas :JSON writable:
         obj = obj.__dict__  # Transforme en :dict:
     for el1 in obj:  # Détecte si des éléments de :obj: sont :JSON writable:
             if type(obj[el1]) in NonJSONWritableList or type(el1) in NonJSONWritableList:
                 obj = f2(obj[el1])
+    for el1 in obj:
             if type(obj[el1]) is list:
                 curr_list = obj[el1]
                 new_list = []
@@ -56,9 +50,22 @@ def f2(obj):
                 obj[el1] = new_list.copy()
     return obj
 
-#S2 = S.__copy__()
-S = f2(S)
+errors = 0
+done = 0
+for _ in np.arange(1000):
+    S = System()
+    S.createPlanet()
+    S2 = S.__copy__()
+    try:
+        S2 = f2(S2)
+        done += 1
+    except:
+        errors += 1
+        done += 1
+print(f"done {done} times with {errors} errors")
+
 #SaveInJSON(S2,path + "S1")
+
 
 """
     if type(obj) in NonJSONWritableList:  # Verifie si :obj: est pas :JSON writable:
